NPC and Monster Range Based Spawns work, Character saving works, LP Regenerates. Lots of work went into this :)

+++ Server/GameServer/NtlPacketGU.h
@@ -1459,8 +1459,8 @@ END_PROTOCOL()*/
 BEGIN_PROTOCOL(GU_UPDATE_CHAR_LP)
 	HOBJECT			handle;
 	DWORD			dwLpEpEventId;
-	WORD			wCurLP;
-	WORD			wMaxLP;
+	DWORD			wCurLP;
+	DWORD			wMaxLP;
 END_PROTOCOL()
 //------------------------------------------------------------------
 BEGIN_PROTOCOL(GU_UPDATE_CHAR_EP)
@@ -1480,8 +1480,8 @@ END_PROTOCOL()
 BEGIN_PROTOCOL(GU_UPDATE_CHAR_LP_EP)
 	HOBJECT			handle;
 	DWORD			dwLpEpEventId;
-	WORD			wCurLP;
-	WORD			wMaxLP;
+	DWORD			wCurLP;
+	DWORD			wMaxLP;
 	WORD			wCurEP;
 	WORD			wMaxEP;
 END_PROTOCOL()
